feat(Online store manager): update the code using inquirer
- I use inquirer module to add interactive to CLI commands.
- inquirer make the CLI commands like a form, ask the user then take the answer.
- Use FileSystem to update the data on json file when we add/update/delete.
- json file work like a DB.

+++ Node JS/Day 02/Lab Solution/Solution using CLI/app.js
@@ -1,6 +1,7 @@
-const { Command } = require('commander');
-const { error } = require('console');
-const { response } = require('express');
+import { Command } from 'commander';
+import { body } from 'express-validator';
+import inquirer from 'inquirer';
+import { type } from 'os';
 
 const program = new Command();
 
@@ -15,27 +16,47 @@ program
     .command('list')
     .alias('l')
     .description('List all products in the store.')
-    .action((options) => {
+    .action(() => {
         console.log("Products: ");
         fetch('http://localhost:8011/')
             .then((response) => response.json())
             .then((products)=> console.log(products));
     })
 
+    
+const addNewProductQuetions = [
+    {
+        type: 'input',
+        name: "name",
+        message: "What's the product name?",
+    },
+    {
+        type: 'number',
+        name: 'price',
+        message: "What's the product price?",
+    }
+];
 program
     .command('add')
     .alias('a')
     .description('Add new product to the store.')
-    .argument('<name>', 'add product name')
-    .argument('<price>', 'add product price')
-    .action((name, price) => {
-        fetch('http://localhost:8011/addProduct', {
-            method: "POST",
-            body: JSON.stringify({name, price: +price}),
-            headers: {"Content-type": "application/json"}
-        }).then((response) => response.json())
-        .then(out => console.log(out))
-        .catch(err => console.log(err));
+    .action(() => {
+        inquirer
+            .prompt(addNewProductQuetions)
+            .then((answers) => {
+                if(answers.name === undefined || answers.price === undefined) {
+                    console.log("You don't enter product name or product price or both.");
+                    return;
+                }
+
+                fetch('http://localhost:8011/addProduct', {
+                    method: "POST",
+                    body: JSON.stringify({name:answers.name , price: +answers.price}),
+                    headers: {"Content-type": "application/json"}
+                }).then((response) => response.json())
+                .then(out => console.log(out))
+                .catch(err => console.log(err));
+            })
     })
 
 
@@ -62,19 +83,38 @@ program
         .catch(err => console.log(err));
     })
 
-
+const deleteProductQuestions = [
+    {
+        type: 'number',
+        name: "id",
+        message: "What's the product id that you want to delete it?",
+    },
+    {
+        type: 'confirm',
+        name: "delete",
+        message: "Are you sure to delete it?",
+    }
+];
 program
     .command('delete')
     .alias('d')
     .description('delete a product from the store using productId.')
-    .argument('<id>', 'productId that will be deleted')
-    .action(id => {
-        fetch(`http://localhost:8011/deleteProduct/${id}`, {
-            method: "DELETE",
-        })
-        .then(response => response.json())
-        .then(out => console.log(out))
-        .catch(err => console.log(err));
+    .action(() => {
+        inquirer
+            .prompt(deleteProductQuestions)
+            .then(answers => {
+                if(answers.delete) {
+                    fetch(`http://localhost:8011/deleteProduct/${answers.id}`, {
+                        method: "DELETE",
+                    })
+                    .then(response => response.json())
+                    .then(out => console.log(out))
+                    .catch(err => console.log(err));
+                } else {
+                    console.log("Product doesn't deleted.")
+                }
+            })
+        
     })
 
 program.parse();
